ddd, Aug 8/16/2022, YYYY, 12:06 A +0

+++ CS50P/fuel/fuel.py
@@ -1,6 +1,6 @@
 def main() :
     num = get_int("Fraction: ")
-    newNum = num * 100
+    newNum = int(num * 100)
 
     if newNum <= 1 :
         print("E")
